An ActTrust actigraphy record

A small sample (1 day length) of an ActTrust actigraphy record.

Note: We're always looking for actigraphy data samples. If you have a personal actigraphy record made with a device that is not listed in the actverse package and is ok with sharing it, please contact the package authors.

ActTrust is a trademark of Condor Instruments Ltda.

acttrust

Format

A tsibble with 17 columns and 1,441 rows:

timestamp

Date and time of the recorded events.

R class: POSIXct.

pim

Wrist activity in Proportional Integral Mode (PIM).

R class: numeric.

tat

Wrist activity in Time Above Threshold (TAT) mode.

R class: numeric.

zcm

Wrist activity in Zero Crossing Mode (ZCM).

R class: numeric.

orientation

Device orientation.

R class: numeric.

wrist_temperature

Wrist temperature in degrees Celsius.

R class: numeric.

external_temperature

External temperature in degrees Celsius (unreliable data).

R class: numeric.

light

Light intensity in lux.

R class: numeric.

ambient_light

Ambient light intensity in microwatts/square meter.

R class: numeric.

red_light

Red light intensity in microwatts/square centimeter.

R class: numeric.

green_light

Green light intensity in microwatts/square centimeter.

R class: numeric.

blue_light

Blue light intensity in microwatts/square centimeter.

R class: numeric.

ir_light

Infrared light intensity in microwatts/square centimeter.

R class: numeric.

uva_light

Ultraviolet A light intensity in microwatts/square centimeter.

R class: numeric.

uvb_light

Ultraviolet B light intensity in microwatts/square centimeter.

R class: numeric.

event

A binary number indicating the moments when the subject pressed the event button: 0 when the button is not pressed and 1 when the button is pressed. This is commonly used to indicate when the subject goes to sleep or wakes up.

R class: numeric.

state

An integer number indicating the subject state: 0 for awake, 1 for sleeping, 2 for resting, 4 for offwrist (device removal), 6 for editable state 1, 7 for editable state 2, and 8 for editable state 3. These states can be assigned by a specialist or by a validated inference algorithm.

R class: numeric.

Source

Created by Daniel Vartanian (package author).

Details

This dataset was created by one of the actverse package authors and is free for all kinds of use. You can download the full dataset at https://zenodo.org/record/4898822.

Data wrangling

This data was tidied and validated with the default settings of read_acttrust().

The process of tiding a dataset is understood as transforming it in input data, like described in Loo and Jonge (2018). It's a very similar process of tiding data described in the workflow proposed by Wickham and Grolemund (n.d.).

The process of validating a dataset is understood as detecting invalid data, by checking whether data satisfies certain assumptions from domain knowledge, to then, removing or, if possible, fixing them. This process can be considered as part of the process of transforming data, described in the workflow proposed by Wickham and Grolemund (n.d.).

To learn more about the concept of tidy data, see Wickham (2014) and Wickham and Grolemund (n.d.). You can find more about data validation and error location in Loo and Jonge (2018).

Variable naming

The actverse package stick to the tidyverse principles. Our code and naming schemes follow the guidelines of the tidyverse style guide.

Variable classes

R has a vast number of time series standards. We choose to work with tsibble because we think is the best standard that follows the tidyverse principles.

To make sure that the resulting data from actverse could be easily transformed to other standards, we decided to maintain all variable classes as numeric, except for the index (class POSIXct).

If you're looking for a good way to easily transform your data to other time series standards, check the tsbox package.

Raw data

The raw data for this dataset can be found with raw_data("acttrust.txt").

Metadata

Subject

Age: 35.

Gender: Male.

Height: 180 cm.

Weight: 85 kg.

BMI: ~26.235 (overweight).

Record

Body part: Left wrist (non-dominant wrist).

Start: 2021-04-24 04:14:25.

End: 2021-05-31 03:39:06.

Days: 37

Epoch (sample interval): 60 seconds.

Mode: PIM, TAT, and ZCM.

TAT Threshold: 1024.

Location: Sao Paulo, SP, Brazil.

UTC: -3.

DST: FALSE.

Light phase: 06:23:59 - 17:44:58.

Dark phase: 17:44:59 - 06:23:58.

Actigraph

Brand: Condor Instruments.

Model: ActTrust 1.

Hardware version: 3.1.

Firmware version: 3.7.

Manual: 1.0.13 (PT-BR).

Software: ActStudio Alpha 1.0.17 (Win | Mac).

Scoring settings

Scoring algorithm: Condor Instruments.

The sleep diary data was not used for scoring.

Scoring review

Offwrist periods were indicated from sudden drops in wrist temperature.

Micro sleep states (isolated states lasting less than 5 minutes) were removed.

Minor adjustments were made on the sleep onset and sleep offset, taking into consideration the event button and sleep diary records.

Sleep statistics settings

Multiple main sleep periods: TRUE.

Main sleep period time of day: Day or Night.

Minimum duration for main sleep periods: 01:00:00.

Awake period limit: 02:00:00.

References

Condor Instruments. (2019). ActTrust - user manual. Sao Paulo, Brazil. Model AT0503. ActStudio version 1.0.13.

Van der Loo, M., & De Jonge, E. (2018). Statistical data cleaning with applications in R. Hooboken, NJ: John Wiley & Sons. doi:10.1002/9781118897126 .

Wickham, H. (2014). Tidy Data. Journal of Statistical Software, 59(10), 1-23. doi:10.18637/jss.v059.i10 .

Wickham, H., & Grolemund, G. (n.d.). R for data science. Sebastopol, CA: O'Reilly Media. https://r4ds.had.co.nz