Get sun related statistics from different APIs

get_sun_stats() retrieves sun statistics using different models and data sources.

At the moment, none of the get_sun_stats() methods uses real-world data. All of them are based on models.

Each API have its peculiarities. We recommend checking the API documentation for a better understanding of the mechanisms behind them (see the Details section).

Usage

get_sun_stats(
  latitude,
  longitude,
  date = Sys.Date(),
  tz = "UTC",
  method = "suncalc"
)

Arguments

latitude

A number indicating the latitude of the desired location in decimal degrees.

longitude

A number indicating the longitude of the desired location in decimal degrees.

date

(optional) A Date value indicating the moment in time (default: Sys.Date()).

tz

(optional) A string indicating the time zone of the results. See timezones to learn more (default: "UTC").

method

(optional) A string indicating which method or API to use. Valid values are: "suncalc" and "sunrise-sunset.org". See the Details section to learn more (default: "suncalc").

Value

A list object with the following elements:

date: A Date vector with the same value of the date parameter.

latitude: A number with the same value of the latitude parameter.

longitude: A number with the same value of the longitude parameter.

tz: A string with the same value of the tz parameter.

sunrise_start: An hms value indicating the moment when the top edge of the sun appears on the horizon.

sunrise_end: An hms value indicating the moment when bottom edge of the sun touches the horizon.

golden_hour_end: An hms value indicating the moment when the morning golden hour (soft light, best time for photography) ends.

solar_noon: An hms value indicating the moment when sun is in the highest position.

golden_hour_start: An hms value indicating the moment when the evening golden hour (soft light, best time for photography) starts.

sunset_start: An hms value indicating the moment when the bottom edge of the sun touches the horizon.

sunset_end: An hms value indicating the moment when the sun disappears below the horizon. This is also the moment when the evening civil twilight starts.

dusk: An hms value indicating the moment when the dusk starts. This is also the moment when the evening nautical twilight starts.

nautical_dusk: An hms value indicating the moment when nautical dusk starts. This is also the moment when the evening astronomical twilight starts.

night_start: An hms value indicating the moment when the night starts (dark enough for astronomical observations).

nadir: An hms value indicating the darkest moment of the night, i.e., when the sun is in the lowest position.

night_end: An hms value indicating the moment when the night ends. This is also the moment when the morning astronomical twilight starts.

nautical_dawn: An hms value indicating the moment when nautical dawn starts. This is also the moment when the morning nautical twilight starts.

dawn: An hms value indicating the moment when the dawn starts. This is also the moment when the morning nautical twilight ends and the morning civil twilight starts.

Details

methods argument

At the moment, get_sun_stats() can access the results of two APIs, described below.

"suncalc": Use the sun statistics provided by the suncalc package.

"sunrise-sunset.org": Use the sun statistics provided by the Sunrise-Sunset API (requires an internet connection). Click here to learn more.

The "sunrise-sunset.org" method tends to give a close, but usually lower, result when compared with the "suncalc" method.

Please note that when using method = "sunrise-sunset.org" you need to show attribution with a link to https://sunrise-sunset.org. Also note that summer time adjustments are not included in the returned data when using this method.

Other statistics

The purpose of this function is to return basic statistics about the sun. If you need other related statistics, we recommend checking the following packages.

cptec: An Interface to the CPTEC/ INPE API.

nasapower: NASA POWER API Client.

rnoaa: NOAA Weather Data from R.

See also

Examples

latitude <- -23.5489
longitude <- -46.6388
date <- Sys.Date()
tz <- "America/Sao_Paulo"

get_sun_stats(
  latitude = latitude,
  longitude = longitude,
  date = date,
  tz = tz,
  method = "suncalc"
)
#> $date
#> [1] "2025-11-05"
#> 
#> $latitude
#> [1] -23.5489
#> 
#> $longitude
#> [1] -46.6388
#> 
#> $tz
#> [1] "America/Sao_Paulo"
#> 
#> $sunrise_start
#> 05:19:24
#> 
#> $sunrise_end
#> 05:21:50
#> 
#> $golden_hour_end
#> 05:50:26
#> 
#> $solar_noon
#> 11:51:26
#> 
#> $golden_hour_start
#> 17:52:27
#> 
#> $sunset_start
#> 18:21:03
#> 
#> $sunset_end
#> 18:23:29
#> 
#> $dusk
#> 18:47:18
#> 
#> $nautical_dusk
#> 19:15:29
#> 
#> $night_start
#> 19:44:23
#> 
#> $nadir
#> 23:51:26
#> 
#> $night_end
#> 03:58:30
#> 
#> $nautical_dawn
#> 04:27:24
#> 
#> $dawn
#> 04:55:34
#> 

library(curl)

if (has_internet()) {
  get_sun_stats(
    latitude = latitude,
    longitude = longitude,
    date = date,
    tz = tz,
    method = "sunrise-sunset.org"
  )
}
#> $date
#> [1] "2025-11-05"
#> 
#> $latitude
#> [1] -23.5489
#> 
#> $longitude
#> [1] -46.6388
#> 
#> $tz
#> [1] "America/Sao_Paulo"
#> 
#> $sunrise_start
#> 05:10:09
#> 
#> $sunrise_end
#> NA
#> 
#> $golden_hour_end
#> NA
#> 
#> $solar_noon
#> 11:43:41
#> 
#> $golden_hour_start
#> NA
#> 
#> $sunset_start
#> NA
#> 
#> $sunset_end
#> 18:17:13
#> 
#> $dusk
#> 18:39:49
#> 
#> $nautical_dusk
#> 19:08:01
#> 
#> $night_start
#> 19:36:58
#> 
#> $nadir
#> NA
#> 
#> $night_end
#> 03:50:24
#> 
#> $nautical_dawn
#> 04:19:20
#> 
#> $dawn
#> 04:47:33
#>