Get data from a Zenodo record

get_from_zenodo() allows you to easy download a Zenodo record or a Zenodo file.

This function only works for Zenodo created DOIs (not when the DOI is, for example, derived from Zookeys).

get_from_zenodo(doi, path = ".", file = NULL, parallel = FALSE)

Arguments

doi

A string indicating a Zenodo DOI (Digital Object Identifier) starting with "10.5281/zenodo.". See the Examples section to learn more.

path

(optional) a string indicating a directory path where the data must be downloaded (default: ".").

file

(optional) a character object with the names of the files that must be downloaded. If NULL, the function will download the entire record (default: NULL).

parallel

(optional) a logical value indicating if the function must run a number of parallel processes, each downloading another file (requires the parallel package). This is useful when multiple large files are present in the Zenodo record, which otherwise would be downloaded sequentially. Of course, the operation is limited by bandwidth and traffic limitations (default: FALSE).

Value

A character object with the paths for all the files downloaded.

Details

Requirements

This function requires an internet connection and the curl, jsonlite, parallel, and tools packages to work.

If you don't have any or one of the packages mentioned above, you can install them with install.packages("curl", "jsonlite", "parallel", "tools").

Zenodo API

You can find more about the Zenodo API at https://developers.zenodo.org/.

License

get_from_zenodo() code is based on the download_zenodo() function found in the inborutils package of the Research Institute for Nature and Forest (INBO). download_zenodo() was created by Hans Van Calster (hans.vancalster@inbo.be) and Floris Vanderhaeghe (floris.vanderhaeghe@inbo.be).

We give our thanks for the INBO institute and for all developers involved in this piece of software.

Please note that this code comes with an MIT License. You can read the latter below.

Copyright (c) 2016 Instituut voor Natuur en Bosonderzoek (INBO)

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in
all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.

See also

Examples

# \dontrun{
## Downloading a single file from a Zenodo record

path <- tempdir()
file <- "sleep-diary.txt"

get_from_zenodo(
    doi = "10.5281/zenodo.4898822", path = path, file = file
    )
#> ℹ Downloading metadata
#> ✔ Downloading metadata [960ms]
#> 
#> ℹ Downloading file
#> ✔ Downloading file [13ms]
#> 
#> ℹ Checking file integrity
#> ✔ Checking file integrity [10ms]
#> 
#> [1] "/tmp/RtmpUamPS7/sleep-diary.txt"

readLines(file.path(path, file))
#>  [1] "BED TIME;GET UP TIME"                   
#>  [2] "24/04/2021 10:15:00;24/04/2021 19:40:00"
#>  [3] "25/04/2021 16:15:00;25/04/2021 23:00:00"
#>  [4] "26/04/2021 18:00:00;27/04/2021 01:30:00"
#>  [5] "27/04/2021 19:15:00;28/04/2021 02:00:00"
#>  [6] "28/04/2021 20:00:00;29/04/2021 05:30:00"
#>  [7] "29/04/2021 22:30:00;30/04/2021 06:30:00"
#>  [8] "30/04/2021 22:00:00;01/05/2021 07:05:00"
#>  [9] "01/05/2021 23:30:00;02/05/2021 08:40:00"
#> [10] "02/05/2021 23:00:00;03/05/2021 08:45:00"
#> [11] "04/05/2021 00:30:00;04/05/2021 09:45:00"
#> [12] "05/05/2021 02:00:00;05/05/2021 10:45:00"
#> [13] "07/05/2021 02:00:00;07/05/2021 12:00:00"
#> [14] "08/05/2021 05:30:00;08/05/2021 13:00:00"
#> [15] "09/05/2021 03:30:00;09/05/2021 12:18:00"
#> [16] "10/05/2021 05:45:00;10/05/2021 12:30:00"
#> [17] "11/05/2021 05:45:00;11/05/2021 13:45:00"
#> [18] "12/05/2021 07:45:00;12/05/2021 13:30:00"
#> [19] "13/05/2021 04:00:00;13/05/2021 10:30:00"
#> [20] "14/05/2021 02:00:00;14/05/2021 09:45:00"
#> [21] "15/05/2021 02:30:00;15/05/2021 12:15:00"
#> [22] "17/05/2021 03:00:00;17/05/2021 13:00:00"
#> [23] "18/05/2021 04:30:00;18/05/2021 11:00:00"
#> [24] "19/05/2021 01:30:00;19/05/2021 11:45:00"
#> [25] "20/05/2021 04:30:00;20/05/2021 13:30:00"
#> [26] "21/05/2021 04:30:00;21/05/2021 13:00:00"
#> [27] "22/05/2021 03:45:00;22/05/2021 13:45:00"
#> [28] "23/05/2021 04:30:00;23/05/2021 13:15:00"
#> [29] "24/05/2021 07:15:00;24/05/2021 15:15:00"
#> [30] "28/05/2021 02:30:00;28/05/2021 13:00:00"

## Downloading all the files from a Zenodo record

get_from_zenodo(
    doi = "10.5281/zenodo.4898822", path = tempdir(), file = NULL
    )
#> ℹ Downloading metadata
#> ✔ Downloading metadata [435ms]
#> 
#> ℹ Downloading files
#> ✔ Downloading files [11ms]
#> 
#> ■■■■■■■■■■                        29% | ETA: 24s
#> ■■■■■■■■■■■■■■■■■■                57% | ETA: 17s
#> ■■■■■■■■■■■■■■■■■■■■■■            71% | ETA: 34s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> 
#> ℹ Checking file integrity
#> ✔ Checking file integrity [73ms]
#> 
#> [1] "/tmp/RtmpUamPS7/actogram.png"        
#> [2] "/tmp/RtmpUamPS7/processed.txt"       
#> [3] "/tmp/RtmpUamPS7/raw.txt"             
#> [4] "/tmp/RtmpUamPS7/raw_report.txt"      
#> [5] "/tmp/RtmpUamPS7/report.pdf"          
#> [6] "/tmp/RtmpUamPS7/sleep-diary.txt"     
#> [7] "/tmp/RtmpUamPS7/sleep_statistics.csv"

## Downloading all the files from a Zenodo record using parallel
## computation

get_from_zenodo(
    doi = "10.5281/zenodo.4898822", path = tempdir(), file = NULL,
    parallel = TRUE
    )
#> ℹ Downloading metadata
#> ✔ Downloading metadata [288ms]
#> 
#> ℹ Initializing parallel download on 7 R session nodes.
#> ℹ Starting parallel downloads
#> ℹ This may take a while. No progress bar will be available. Be patient.
#> ℹ Starting parallel downloads
#> ✔ Starting parallel downloads [1m 6.9s]
#> 
#> ℹ Ending parallel downloads
#> ✔ Ending parallel downloads [87ms]
#> 
#> ℹ Checking file integrity
#> ✔ Checking file integrity [64ms]
#> 
#> [1] "/tmp/RtmpUamPS7/actogram.png"        
#> [2] "/tmp/RtmpUamPS7/processed.txt"       
#> [3] "/tmp/RtmpUamPS7/raw.txt"             
#> [4] "/tmp/RtmpUamPS7/raw_report.txt"      
#> [5] "/tmp/RtmpUamPS7/report.pdf"          
#> [6] "/tmp/RtmpUamPS7/sleep-diary.txt"     
#> [7] "/tmp/RtmpUamPS7/sleep_statistics.csv"
# }